проба встраивания в нзб в дкп

+++ app/src/main/java/com/example/matek3022/nirs/Pixel.kt
@@ -1,8 +1,8 @@
 package com.example.matek3022.nirs
 
-data class Pixel(var red: Float, var green: Float, var blue: Float) {
+data class Pixel(var red: Int, var green: Int, var blue: Int) {
     fun compare(pixel: Pixel) = Math.abs(red - pixel.red) + Math.abs(green - pixel.green) + Math.abs(blue - pixel.blue)
-    fun maxDiff(pixel: Pixel): Float{
+    fun maxDiff(pixel: Pixel): Int{
         val a1 = Math.abs(red - pixel.red)
         val a2 = Math.abs(green - pixel.green)
         val a3 = Math.abs(blue - pixel.blue)
